refactor: extract proxy dialing into proxy_dial.go
Separate proxy tunneling (SOCKS5, HTTP/HTTPS CONNECT, env fallback)
from TLS fingerprinting in utls_transport.go. Introduces ProxyDialer
as a self-contained proxy dialing abstraction.

- utls_transport.go: 277 → 113 lines (TLS fingerprint + client cache only)
- proxy_dial.go: 202 lines (all proxy tunneling logic)
- Public API unchanged (NewAnthropicHttpClient signature preserved)
